tests/requirements.txt
# Hyperliquid SDK
hyperliquid-python-sdk>=0.19.0

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Ethereum utilities
eth-account>=0.9.0
web3>=6.11.0

# JSON and data handling
msgpack>=1.0.0

# Async HTTP client
aiohttp>=3.9.0

# Additional testing utilities
pytest-mock>=3.12.0
pytest-timeout>=2.2.0